Refactor api responses
