--- src/main/kotlin/ru/neexol/models/LessonTime.kt
@@ -1,7 +1,0 @@
-package ru.neexol.models
-
-@kotlinx.serialization.Serializable
-data class LessonTime(
-    val begin: String,
-    val end: String
-)
