--- src/main/kotlin/ru/neexol/models/Schedule.kt
@@ -1,8 +1,0 @@
-package ru.neexol.models
-
-@kotlinx.serialization.Serializable
-data class Schedule(
-    val group: String,
-    val checksum: String,
-    val lessons: List<Lesson>
-)
