+++ src/main/kotlin/ru/neexol/models/responses/LessonTimeResponse.kt
@@ -1,0 +1,19 @@
+package ru.neexol.models.responses
+
+@kotlinx.serialization.Serializable
+data class LessonTimeResponse(
+    val begin: String,
+    val end: String
+)
+
+val lessonTimes = listOf(
+    LessonTimeResponse("09:00", "10:30"),
+    LessonTimeResponse("10:40", "12:10"),
+    LessonTimeResponse("12:40", "14:10"),
+    LessonTimeResponse("14:20", "15:50"),
+    LessonTimeResponse("16:20", "17:50"),
+    LessonTimeResponse("18:00", "19:30"),
+    LessonTimeResponse("19:40", "21:10"),
+    LessonTimeResponse("18:30", "20:00"),
+    LessonTimeResponse("20:10", "21:40")
+)
