+++ src/main/kotlin/ru/neexol/models/responses/ScheduleResponse.kt
@@ -1,0 +1,8 @@
+package ru.neexol.models.responses
+
+@kotlinx.serialization.Serializable
+data class ScheduleResponse(
+    val group: String,
+    val checksum: String,
+    val lessons: List<LessonResponse>
+)
