+++ src/main/kotlin/ru/neexol/routes/api/timesRoute.kt
@@ -3,7 +3,7 @@ package ru.neexol.routes.api
 import io.ktor.server.application.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
-import ru.neexol.utils.lessonTimes
+import ru.neexol.models.responses.lessonTimes
 
 fun Route.timesRoute() {
     route("times") {
